+++ packages/lucia-auth/src/auth/index.ts
@@ -1,31 +1,15 @@
-import type { Env, User } from "../types.js";
-import {
-	authenticateUserFunction,
-	createUserFunction,
-	deleteUserFunction,
-	getUserFunction,
-	getUserByProviderIdFunction,
-	updateUserAttributesFunction,
-	updateUserPasswordFunction,
-	updateUserProviderIdFunction
-} from "./user/index.js";
-import { validateRequestHeadersFunction } from "./request.js";
-import {
-	createSessionFunction,
-	deleteDeadUserSessionsFunction,
-	renewSessionFunction,
-	invalidateAllUserSessionsFunction,
-	validateSessionFunction,
-	validateSessionUserFunction,
-	invalidateSessionFunction,
-	getSessionFunction,
-	generateSessionIdFunction,
-	getSessionUserFunction
-} from "./session.js";
-import { createSessionCookiesFunction } from "./cookie.js";
+import type { Env, MinimalRequest, Session, SessionSchema, User } from "../types.js";
+import { Cookie, CookieOption, createSessionCookie } from "./cookie.js";
 import { Adapter, SessionAdapter, UserAdapter, UserData, UserSchema } from "../types.js";
 import { logError } from "../utils/log.js";
-import { generateHashWithScrypt, validateHash } from "../utils/crypto.js";
+import {
+	generateHashWithScrypt,
+	generateRandomString,
+	validateScryptHash
+} from "../utils/crypto.js";
+import { LuciaError } from "./error.js";
+import { parseCookie } from "../utils/cookie.js";
+import { getSessionFromDatabaseData } from "./session.js";
 
 export { SESSION_COOKIE_NAME } from "./cookie.js";
 
@@ -34,18 +18,35 @@ export const lucia = <C extends Configurations>(configs: C) => {
 };
 
 const validateConfigurations = (configs: Configurations) => {
-	const isAdapterIdentified = !configs.adapter;
-	if (isAdapterIdentified) {
+	const isAdapterProvided = configs.adapter;
+	if (!isAdapterProvided) {
 		logError('Adapter is not defined in configuration ("config.adapter")');
 		process.exit(1);
 	}
 };
 
 export class Auth<C extends Configurations = any> {
-	public configs: UserConfig<C>;
+	private adapter: Adapter;
+	private generateUserId: () => MaybePromise<string | null>;
+	private sessionCookie: CookieOption[];
+	private sessionTimeout: {
+		activePeriod: number;
+		idlePeriod: number;
+	};
+	private env: Env;
+	private hash: {
+		generate: (s: string) => MaybePromise<string>;
+		validate: (s: string, hash: string) => MaybePromise<boolean>;
+	};
+	private autoDatabaseCleanup: boolean;
+	private transformUserData: (
+		userData: UserSchema
+	) => C["transformUserData"] extends {} ? ReturnType<C["transformUserData"]> : { userId: string };
+	private csrfProtection: boolean;
+
 	constructor(configs: C) {
 		validateConfigurations(configs);
-		const defaultCookieOption: CookieOption = {
+		const defaultSessionCookieOption: CookieOption = {
 			sameSite: "lax",
 			path: "/"
 		};
@@ -57,81 +58,267 @@ export class Auth<C extends Configurations = any> {
 				delete configs.adapter.session.getSessionAndUserBySessionId;
 			}
 		}
-		const adapter =
+		this.adapter =
 			"user" in configs.adapter
 				? { ...configs.adapter.user, ...configs.adapter.session }
 				: configs.adapter;
-		this.configs = {
-			adapter,
-			generateCustomUserId: configs.generateCustomUserId || (async () => null),
-			env: configs.env,
-			csrfProtection: configs.csrfProtection || true,
-			sessionTimeout: configs.sessionTimeout || 1000 * 60 * 60 * 24,
-			idlePeriodTimeout: configs.idlePeriodTimeout || 1000 * 60 * 60 * 24 * 14,
-			transformUserData: ({ id, hashed_password, provider_id, ...attributes }) => {
-				const transform =
-					configs.transformUserData ||
-					(({ id }) => {
-						return {
-							userId: id
-						};
-					});
-				return transform({ id, ...attributes }) as User;
-			},
-			sessionCookieOptions: configs.sessionCookieOptions || [defaultCookieOption],
-			deleteCookieOptions: configs.deleteCookieOptions || [],
-			hash: {
-				generate: configs.hash?.generate ?? generateHashWithScrypt,
-				validate: configs.hash?.validate ?? validateHash
-			}
+		this.generateUserId = configs.generateCustomUserId ?? (async () => null);
+		this.env = configs.env;
+		this.csrfProtection = configs.csrfProtection ?? true;
+		this.sessionTimeout = {
+			activePeriod: configs.sessionTimeout?.activePeriod ?? 1000 * 60 * 60 * 24,
+			idlePeriod: configs.sessionTimeout?.idlePeriod ?? 1000 * 60 * 60 * 24 * 14
 		};
-		this.getUser = getUserFunction(this);
-		this.getUserByProviderId = getUserByProviderIdFunction(this);
-		this.getSessionUser = getSessionUserFunction(this);
-		this.createUser = createUserFunction(this);
-		this.updateUserAttributes = updateUserAttributesFunction(this);
-		this.updateUserProviderId = updateUserProviderIdFunction(this);
-		this.updateUserPassword = updateUserPasswordFunction(this);
-		this.deleteUser = deleteUserFunction(this);
-		this.authenticateUser = authenticateUserFunction(this);
-
-		this.getSession = getSessionFunction(this);
-		this.validateSession = validateSessionFunction(this);
-		this.validateSessionUser = validateSessionUserFunction(this);
-		this.generateSessionId = generateSessionIdFunction(this);
-		this.createSession = createSessionFunction(this);
-		this.renewSession = renewSessionFunction(this);
-		this.invalidateSession = invalidateSessionFunction(this);
-		this.invalidateAllUserSessions = invalidateAllUserSessionsFunction(this);
-		this.deleteDeadUserSessions = deleteDeadUserSessionsFunction(this);
-
-		this.validateRequestHeaders = validateRequestHeadersFunction(this);
-
-		this.createSessionCookies = createSessionCookiesFunction(this);
+		this.autoDatabaseCleanup = configs.autoDatabaseCleanup ?? true;
+		this.transformUserData = ({ id, hashed_password, provider_id, ...attributes }) => {
+			const transform =
+				configs.transformUserData ??
+				(({ id }) => {
+					return {
+						userId: id
+					};
+				});
+			return transform({ id, ...attributes }) as User;
+		};
+		this.sessionCookie = configs.sessionCookie ?? [defaultSessionCookieOption];
+		this.hash = {
+			generate: configs.hash?.generate ?? generateHashWithScrypt,
+			validate: configs.hash?.validate ?? validateScryptHash
+		};
 	}
-	public getUser: ReturnType<typeof getUserFunction>;
-	public getUserByProviderId: ReturnType<typeof getUserByProviderIdFunction>;
-	public getSessionUser: ReturnType<typeof getSessionUserFunction>;
-	public createUser: ReturnType<typeof createUserFunction>;
-	public updateUserAttributes: ReturnType<typeof updateUserAttributesFunction>;
-	public updateUserProviderId: ReturnType<typeof updateUserProviderIdFunction>;
-	public updateUserPassword: ReturnType<typeof updateUserPasswordFunction>;
-	public deleteUser: ReturnType<typeof deleteUserFunction>;
-	public authenticateUser: ReturnType<typeof authenticateUserFunction>;
 
-	public getSession: ReturnType<typeof getSessionFunction>;
-	public validateSession: ReturnType<typeof validateSessionFunction>;
-	public validateSessionUser: ReturnType<typeof validateSessionUserFunction>;
-	public generateSessionId: ReturnType<typeof generateSessionIdFunction>;
-	public createSession: ReturnType<typeof createSessionFunction>;
-	public renewSession: ReturnType<typeof renewSessionFunction>;
-	public invalidateSession: ReturnType<typeof invalidateSessionFunction>;
-	public invalidateAllUserSessions: ReturnType<typeof invalidateAllUserSessionsFunction>;
-	public deleteDeadUserSessions: ReturnType<typeof deleteDeadUserSessionsFunction>;
-
-	public validateRequestHeaders: ReturnType<typeof validateRequestHeadersFunction>;
-
-	public createSessionCookies: ReturnType<typeof createSessionCookiesFunction>;
+	public getUser = async (userId: string): Promise<User> => {
+		const databaseUser = await this.adapter.getUser(userId);
+		if (!databaseUser) throw new LuciaError("AUTH_INVALID_USER_ID");
+		const user = this.transformUserData(databaseUser);
+		return user;
+	};
+	public getUserByProviderId = async (provider: string, identifier: string): Promise<User> => {
+		const providerId = `${provider}:${identifier}`;
+		const databaseUser = await this.adapter.getUserByProviderId(providerId);
+		if (!databaseUser) throw new LuciaError("AUTH_INVALID_PROVIDER_ID");
+		const user = this.transformUserData(databaseUser);
+		return user;
+	};
+	public getSessionUser = async (
+		sessionId: string
+	): Promise<{
+		user: User;
+		session: Session;
+	}> => {
+		if (sessionId.length !== 40) throw new LuciaError("AUTH_INVALID_SESSION_ID");
+		let userData: UserSchema | null;
+		let sessionData: SessionSchema | null;
+		if (this.adapter.getSessionAndUserBySessionId !== undefined) {
+			const databaseUserSession = await this.adapter.getSessionAndUserBySessionId(sessionId);
+			if (!databaseUserSession) throw new LuciaError("AUTH_INVALID_SESSION_ID");
+			userData = databaseUserSession.user;
+			sessionData = databaseUserSession.session;
+		} else {
+			sessionData = await this.adapter.getSession(sessionId);
+			userData = sessionData ? await this.adapter.getUser(sessionData.user_id) : null;
+		}
+		if (!sessionData) throw new LuciaError("AUTH_INVALID_SESSION_ID");
+		const session = getSessionFromDatabaseData(sessionData);
+		if (!session) {
+			if (this.autoDatabaseCleanup) {
+				await this.adapter.deleteSession(sessionId);
+			}
+			throw new LuciaError("AUTH_INVALID_SESSION_ID");
+		}
+		if (!userData) throw new LuciaError("AUTH_INVALID_USER_ID");
+		return {
+			user: this.transformUserData(userData),
+			session
+		};
+	};
+	public createUser = async (
+		provider: string,
+		identifier: string,
+		options?: {
+			password?: string;
+			attributes?: Lucia.UserAttributes;
+		}
+	): Promise<User> => {
+		const providerId = `${provider}:${identifier}`;
+		const attributes = options?.attributes ?? {};
+		const userId = await this.generateUserId();
+		const hashedPassword = options?.password ? await this.hash.generate(options.password) : null;
+		const userData = await this.adapter.setUser(userId, {
+			providerId,
+			hashedPassword: hashedPassword,
+			attributes
+		});
+		const user = this.transformUserData(userData);
+		return user;
+	};
+	public updateUserAttributes = async (
+		userId: string,
+		attributes: Partial<Lucia.UserAttributes>
+	): Promise<User> => {
+		const [userData] = await Promise.all([
+			this.adapter.updateUser(userId, {
+				attributes
+			}),
+			this.autoDatabaseCleanup ? await this.deleteDeadUserSessions(userId) : null
+		]);
+		const user = this.transformUserData(userData);
+		return user;
+	};
+	public updateUserProviderId = async (
+		userId: string,
+		provider: string,
+		identifier: string
+	): Promise<User> => {
+		const providerId = `${provider}:${identifier}`;
+		const [userData] = await Promise.all([
+			this.adapter.updateUser(userId, {
+				providerId
+			}),
+			this.autoDatabaseCleanup ? await this.deleteDeadUserSessions(userId) : null
+		]);
+		const user = this.transformUserData(userData);
+		return user;
+	};
+	public updateUserPassword = async (userId: string, password: string): Promise<User> => {
+		const hashedPassword = password ? await this.hash.generate(password) : null;
+		const [userData] = await Promise.all([
+			this.adapter.updateUser(userId, {
+				hashedPassword
+			}),
+			this.invalidateAllUserSessions(userId)
+		]);
+		const user = this.transformUserData(userData);
+		return user;
+	};
+	public deleteUser = async (userId: string): Promise<void> => {
+		await this.adapter.deleteSessionsByUserId(userId);
+		await this.adapter.deleteUser(userId);
+	};
+	public authenticateUser = async (
+		provider: string,
+		identifier: string,
+		password: string
+	): Promise<User> => {
+		const providerId = `${provider}:${identifier}`;
+		const databaseData = await this.adapter.getUserByProviderId(providerId);
+		if (!databaseData) throw new LuciaError("AUTH_INVALID_PROVIDER_ID");
+		if (!databaseData.hashed_password) throw new LuciaError("AUTH_INVALID_PASSWORD");
+		if (databaseData.hashed_password.startsWith("$2a"))
+			throw new LuciaError("AUTH_OUTDATED_PASSWORD");
+		const isValid = await this.hash.validate(password, databaseData.hashed_password);
+		if (!isValid) throw new LuciaError("AUTH_INVALID_PASSWORD");
+		const user = this.transformUserData(databaseData);
+		return user;
+	};
+	public getSession = async (sessionId: string): Promise<Session> => {
+		if (sessionId.length !== 40) throw new LuciaError("AUTH_INVALID_SESSION_ID");
+		const databaseSession = await this.adapter.getSession(sessionId);
+		if (!databaseSession) throw new LuciaError("AUTH_INVALID_SESSION_ID");
+		const session = getSessionFromDatabaseData(databaseSession);
+		if (!session) {
+			if (this.autoDatabaseCleanup) {
+				await this.adapter.deleteSession(sessionId);
+			}
+			throw new LuciaError("AUTH_INVALID_SESSION_ID");
+		}
+		return session;
+	};
+	public validateSession = async (sessionId: string): Promise<Session> => {
+		const session = await this.getSession(sessionId);
+		if (session.state === "active") return session;
+		const renewedSession = await this.renewSession(sessionId);
+		return renewedSession;
+	};
+	public validateSessionUser = async (
+		sessionId: string
+	): Promise<{ session: Session; user: User }> => {
+		const { session, user } = await this.getSessionUser(sessionId);
+		if (session.state === "active") return { session, user };
+		const renewedSession = await this.renewSession(sessionId);
+		return {
+			session: renewedSession,
+			user
+		};
+	};
+	public generateSessionId = (): [
+		sessionId: string,
+		activePeriodExpires: Date,
+		idlePeriodExpires: Date
+	] => {
+		const sessionId = generateRandomString(40);
+		const activePeriodExpires = new Date(new Date().getTime() + this.sessionTimeout.activePeriod);
+		const idlePeriodExpires = new Date(
+			activePeriodExpires.getTime() + this.sessionTimeout.idlePeriod
+		);
+		return [sessionId, activePeriodExpires, idlePeriodExpires];
+	};
+	public createSession = async (userId: string): Promise<Session> => {
+		const [sessionId, activePeriodExpires, idlePeriodExpires] = this.generateSessionId();
+		await Promise.all([
+			this.adapter.setSession(sessionId, {
+				userId,
+				expires: activePeriodExpires.getTime(),
+				idlePeriodExpires: idlePeriodExpires.getTime()
+			}),
+			this.autoDatabaseCleanup ? await this.deleteDeadUserSessions(userId) : null
+		]);
+		return {
+			userId,
+			activePeriodExpires,
+			sessionId,
+			idlePeriodExpires,
+			state: "active",
+			isFresh: true
+		};
+	};
+	public renewSession = async (sessionId: string): Promise<Session> => {
+		if (sessionId.length !== 40) throw new LuciaError("AUTH_INVALID_SESSION_ID");
+		const sessionData = await this.adapter.getSession(sessionId);
+		if (!sessionData) throw new LuciaError("AUTH_INVALID_SESSION_ID");
+		const session = getSessionFromDatabaseData(sessionData);
+		if (!session) {
+			if (this.autoDatabaseCleanup) {
+				await this.adapter.deleteSession(sessionId);
+			}
+			throw new LuciaError("AUTH_INVALID_SESSION_ID");
+		}
+		const [renewedSession] = await Promise.all([
+			await this.createSession(session.userId),
+			this.autoDatabaseCleanup ? await this.deleteDeadUserSessions(session.userId) : null
+		]);
+		return renewedSession;
+	};
+	public invalidateSession = async (sessionId: string): Promise<void> => {
+		await this.adapter.deleteSession(sessionId);
+	};
+	public invalidateAllUserSessions = async (userId: string): Promise<void> => {
+		await this.adapter.deleteSessionsByUserId(userId);
+	};
+	public deleteDeadUserSessions = async (userId: string): Promise<void> => {
+		const sessions = await this.adapter.getSessionsByUserId(userId);
+		const currentTime = new Date().getTime();
+		const deadSessionIds = sessions
+			.filter((val) => val.idle_expires < currentTime)
+			.map((val) => val.id);
+		if (deadSessionIds.length === 0) return;
+		await this.adapter.deleteSession(...deadSessionIds);
+	};
+	public validateRequestHeaders = (request: MinimalRequest): string => {
+		const cookies = parseCookie(request.headers.get("cookie") ?? "");
+		const sessionId = cookies.auth_session ?? "";
+		const checkForCsrf = request.method !== "GET" && request.method !== "HEAD";
+		if (checkForCsrf && this.csrfProtection) {
+			const origin = request.headers.get("Origin");
+			if (!origin) throw new LuciaError("AUTH_INVALID_REQUEST");
+			const url = new URL(request.url);
+			if (url.origin !== origin) throw new LuciaError("AUTH_INVALID_REQUEST");
+		}
+		return sessionId;
+	};
+	public createSessionCookies = (session: Session | null): Cookie[] => {
+		return this.sessionCookie.map((options) => createSessionCookie(session, this.env, options));
+	};
 }
 
 type MaybePromise<T> = T | Promise<T>;
@@ -146,37 +333,15 @@ interface Configurations {
 	env: Env;
 	generateCustomUserId?: () => Promise<string | null>;
 	csrfProtection?: boolean;
-	sessionTimeout?: number;
-	idlePeriodTimeout?: number;
+	sessionTimeout?: {
+		activePeriod: number;
+		idlePeriod: number;
+	};
 	transformUserData?: (userData: UserData) => Record<string, any>;
-	sessionCookieOptions?: CookieOption[];
-	deleteCookieOptions?: CookieOption[];
+	sessionCookie?: CookieOption[];
 	hash?: {
 		generate: (s: string) => MaybePromise<string>;
 		validate: (s: string, hash: string) => MaybePromise<boolean>;
 	};
+	autoDatabaseCleanup?: boolean;
 }
-
-type CookieOption = {
-	sameSite?: "strict" | "lax";
-	path?: string;
-	domain?: string;
-};
-
-type UserConfig<C extends Configurations = any> = {
-	adapter: Adapter;
-	env: Env;
-	generateCustomUserId: () => Promise<string | null>;
-	csrfProtection: boolean;
-	sessionTimeout: number;
-	idlePeriodTimeout: number;
-	sessionCookieOptions: CookieOption[];
-	deleteCookieOptions: CookieOption[];
-	transformUserData: (
-		userData: UserSchema
-	) => C["transformUserData"] extends {} ? ReturnType<C["transformUserData"]> : { userId: string };
-	hash: {
-		generate: (s: string) => MaybePromise<string>;
-		validate: (s: string, hash: string) => MaybePromise<boolean>;
-	};
-};