+++ packages/lucia-auth/src/auth/user/index.ts
@@ -1,9 +1,0 @@
-export { authenticateUserFunction } from "./authenticate.js";
-export { getUserFunction, getUserByProviderIdFunction } from "./get.js";
-export {
-	createUserFunction,
-	updateUserProviderIdFunction,
-	updateUserAttributesFunction,
-	updateUserPasswordFunction,
-	deleteUserFunction
-} from "./mutate.js";