+++ packages/adapter-mongoose/CHANGELOG.md
@@ -1,13 +1,17 @@
 # CHANGELOG
 
-### v0.1.4
+## 0.1.5
+
+- Update peer dependency
+
+## 0.1.4
 
 - [Fix] Remove `instance of` check for error [#213](https://github.com/pilcrowOnPaper/lucia-auth/issues/213)
 
-### v0.1.3
+## 0.1.3
 
 - Update dependencies
 
-### v0.1.2
+## 0.1.2
 
 - Update peer dependency