+++ packages/adapter-test/CHANGELOG.md
@@ -1,13 +1,17 @@
 # CHANGELOG
 
-### v0.1.4
+## 0.1.5
+
+- Update peer dependency
+
+## 0.1.4
 
 - Fix `testAdapter` test suite so it runs all tests
 
-### v0.1.3
+## 0.1.3
 
 - Update dependencies
 
-### v0.1.2
+## 0.1.2
 
 - Update peer dependency